Move `ConsoleUtils` to `junit-platform-console`
