+++ junit-platform-commons/junit-platform-commons.gradle.kts
@@ -16,8 +16,6 @@ dependencies {
 tasks.jar {
 	val release9ClassesDir = sourceSets.mainRelease9.get().output.classesDirs.singleFile
 	inputs.dir(release9ClassesDir).withPathSensitivity(PathSensitivity.RELATIVE)
-	val release17ClassesDir = sourceSets.mainRelease17.get().output.classesDirs.singleFile
-	inputs.dir(release17ClassesDir).withPathSensitivity(PathSensitivity.RELATIVE)
 	doLast {
 		exec {
 			executable = project.the<JavaToolchainService>().launcherFor(java.toolchain).get()
@@ -26,9 +24,7 @@ tasks.jar {
 				"--update",
 				"--file", archiveFile.get().asFile.absolutePath,
 				"--release", "9",
-				"-C", release9ClassesDir.absolutePath, ".",
-				"--release", "17",
-				"-C", release17ClassesDir.absolutePath, "."
+				"-C", release9ClassesDir.absolutePath, "."
 			)
 		}
 	}
@@ -37,6 +33,5 @@ tasks.jar {
 eclipse {
 	classpath {
 		sourceSets -= project.sourceSets.mainRelease9.get()
-		sourceSets -= project.sourceSets.mainRelease17.get()
 	}
 }
