+++ junit-platform-console/junit-platform-console.gradle.kts
@@ -21,6 +21,8 @@ dependencies {
 
 tasks {
 	shadowJar {
+		val release17ClassesDir = sourceSets.mainRelease17.get().output.classesDirs.singleFile
+		inputs.dir(release17ClassesDir).withPathSensitivity(PathSensitivity.RELATIVE)
 		exclude("META-INF/versions/9/module-info.class")
 		relocate("picocli", "org.junit.platform.console.shadow.picocli")
 		from(projectDir) {
@@ -35,7 +37,9 @@ tasks {
 				args(
 					"--update",
 					"--file", archiveFile.get().asFile.absolutePath,
-					"--main-class", "org.junit.platform.console.ConsoleLauncher"
+					"--main-class", "org.junit.platform.console.ConsoleLauncher",
+					"--release", "17",
+					"-C", release17ClassesDir.absolutePath, "."
 				)
 			}
 		}
@@ -52,5 +56,12 @@ tasks {
 	// So in order to resolve this, it can only run on Java 9
 	osgiProperties {
 		property("-runee", "JavaSE-9")
+	}
+}
+
+eclipse {
+	classpath {
+		sourceSets -= project.sourceSets.mainRelease9.get()
+		sourceSets -= project.sourceSets.mainRelease17.get()
 	}
 }
