+++ junit-platform-console/src/main/java17/org/junit/platform/console/options/ConsoleUtils.java
@@ -8,10 +8,11 @@
  * https://www.eclipse.org/legal/epl-v20.html
  */
 
-package org.junit.platform.commons.util;
+package org.junit.platform.console.options;
 
 import static org.apiguardian.api.API.Status.INTERNAL;
 
+import java.io.Console;
 import java.nio.charset.Charset;
 
 import org.apiguardian.api.API;
@@ -35,6 +36,7 @@ public class ConsoleUtils {
 	 * {@return the charset of the console}
 	 */
 	public static Charset charset() {
-		return Charset.defaultCharset();
+		Console console = System.console();
+		return console != null ? console.charset() : Charset.defaultCharset();
 	}
 }
